+++ ait/commons/util/command/sync.py
@@ -1,29 +1,18 @@
 import os
-import json
-import requests
 from tqdm import tqdm
 from multiprocessing import cpu_count
 from multiprocessing.pool import ThreadPool
-import random
 
 from botocore.exceptions import ClientError
 
 from ait.commons.util.aws_client import Aws
 from ait.commons.util.bucket_policy import new_policy_statement
-from ait.commons.util.common import gen_uuid, format_err, INGEST_UPLOAD_AREA_PREFIX, sizeof_fmt
+from ait.commons.util.common import gen_uuid, format_err, INGEST_UPLOAD_AREA_PREFIX
 from ait.commons.util.local_state import get_selected_area
-
-
-UPLOAD_SERVICE_API = 'https://upload__ENV__archive.data.humancellatlas.org'
+from ait.commons.util.upload_service import notify_upload
 
 
 class CmdSync:
-    """
-    Upload srv steps
-    1. get creds from upload srv api
-    2. upload via awscli or hca-util (set content-type accordingly)
-    3. post notification to upload srv api
-    """
 
     def __init__(self, aws: Aws, args):
         self.aws = aws
@@ -120,48 +109,3 @@ class CmdSync:
 def num_files(ls):
     l = len(ls)
     return f'{l} file{"s" if l > 1 else ""}'
-
-
-"""
-Upload service api endpoints
-POST /v1/area/{upload_area_uuid}/credentials    Create credentials for access to this upload area
-POST /v1/area/{upload_area_uuid}/{filename}     Notify upload of uploaded file
-HEAD /v1/area/{upload_area_uuid}                Check if an upload area exists
-"""
-
-def upload_api_url(env, upload_area_uuid):
-    sub = '.' if env == 'prod' else f'.{env}.'
-    url = UPLOAD_SERVICE_API.replace("__ENV__", sub) + '/v1/area/' + upload_area_uuid
-    return url
-
-
-def create_creds(env, upload_area_uuid):
-    r = requests.post(upload_api_url(env, upload_area_uuid) + '/credentials')
-    if r.status_code == 201:
-        # Created
-        return r.json()
-    #elif r.status_code == 404:
-    #    # Upload area not found
-    #    print(r.json())
-
-
-def notify_upload(env, upload_area_uuid, filename):
-    try:
-        r = requests.post(upload_api_url(env, upload_area_uuid) + '/' + filename)
-        if r.status_code == 202:
-            return True # File upload notification added to queue
-    except requests.exceptions.RequestException as e:
-        pass # silently pass
-    return False # notify failed
-
-
-def check_upload_area_exists(env, upload_area_uuid):
-    try:
-        r = requests.head(upload_api_url(env, upload_area_uuid))
-        if r.status_code == 200:
-            return True
-        else:
-            # Upload area does not exist.
-            return False
-    except requests.exceptions.RequestException as e:
-        print(str(e))
